requirements.txt
# ═══════════════════════════════════════════════════════════
# LangChain Core
# ═══════════════════════════════════════════════════════════
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.30

# ═══════════════════════════════════════════════════════════
# MCP (Model Context Protocol)
# ═══════════════════════════════════════════════════════════
fastmcp>=0.1.0
mcp>=0.1.0
langchain-mcp-adapters>=0.1.0

# ═══════════════════════════════════════════════════════════
# Data Processing
# ═══════════════════════════════════════════════════════════
pandas>=2.0.0
duckdb>=0.9.0

# ═══════════════════════════════════════════════════════════
# Streamlit UI
# ═══════════════════════════════════════════════════════════
streamlit>=1.30.0
Pillow>=10.0.0

# ═══════════════════════════════════════════════════════════
# Testing
# ═══════════════════════════════════════════════════════════
pytest>=7.4.0
pytest-cov>=4.1.0

# ═══════════════════════════════════════════════════════════
# Utilities
# ═══════════════════════════════════════════════════════════
python-dotenv>=1.0.0
tqdm>=4.66.0
tomli>=2.0.0
filelock>=3.12.0